requirements-gpu.txt
tensorflow-gpu==2.1.0
numpy
opencv-python==4.1.1.26
lxml
tqdm
flask
seaborn
pillow
pyautogui